requirements.txt
django>=6.0.3
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter>=24.2
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
django-environ>=0.11.2
django-axes>=6.1.1
django-ratelimit>=4.1.0
django-auditlog>=3.0.0
django-redis>=5.4.0
django-csp>=3.7
django-two-factor-auth[phonenumbers,call,sms]>=1.15.5
celery>=5.3.6
redis>=5.0.1
gunicorn>=21.2.0
reportlab>=4.2.2
google-api-python-client>=2.169.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.2
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-magic>=0.4.27
django-cleanup>=8.0.0
